lots of development related to getting ProjectConfig, DatasetConfig, ProjectRegistry, and DatasetRegistry connected
