+++ dsgrid/config/dataset_config.py
@@ -16,10 +16,12 @@ from pydantic.fields import Field
 from pydantic.class_validators import root_validator, validator
 
 from dsgrid.exceptions import DSGBaseException
-from dsgrid.config.project_config import (DSGBaseModel, Dimension,
-                                          TimeDimension, DimensionType)
-from dsgrid.registry.load_registry import load_project_registry
+from dsgrid.dimension.base import DimensionType, DSGBaseModel
 from dsgrid.utils.utilities import run_command
+
+
+from dsgrid.config._config import (
+    TimeDimension, Dimension, ConfigRegistrationDetails)
 
 
 # TODO: likely needs refinement (missing mappings)
@@ -40,35 +42,15 @@ class DSGDatasetParquetType(Enum):
     PROJECT_DIMENSION_MAPPING = 'project_dimension_mapping'  # optional
 
 
-class DatasetDimension(Dimension):
-    use_project_dimension: Optional[bool] = Field(
-        title="use_project_dimension",
-        description="Boolean to use same dimension settings as the project",
-    )
-
-    @validator("use_project_dimension")
-    def apply_project_dimension_if_true(cls, use_project_dimension, values):
-        print(values)
-
-
-class DatasetDimensions(DSGBaseModel):
-    """Contains dimensions defined by a dataset"""
-    dimensions: List[Union[DatasetDimension, TimeDimension]] = Field(
-        title="dimensions",
-        description="dimensions defined by the dataset",
-    )
-    # TODO: add the same handle_dimension union as project config
-    # TODO: why do I not get the same errors here as were recieved
-    #   at the project config level?
-
 
 # TODO will need to rename this as it really should be more generic inputs
 #   and not just sector inputs. "InputSectorDataset" is already taken in
 #   project_config.py
+# TODO: this already assumes that the data is formatted for DSG, however,
+#       we may want the dataset config to be "before" any DSG parquets get
+#       formatted.
 class InputSectorDataset(DSGBaseModel):
-    # TODO: this already assumes that the data is formatted for DSG, however,
-    #       we may want the dataset config to be "before" any DSG parquets get
-    #       formatted.
+    """Input dataset configuration class"""
     data_type: DSGDatasetParquetType = Field(
         title="data_type",
         alias="type",
@@ -79,22 +61,16 @@ class InputSectorDataset(DSGBaseModel):
         description="directory with parquet files"
     )
 
-    # TODO: validate data matches dimensions;
-    #   must not have any missing dimension value;
-    #   check expected counts of things?
-    #   must not have any missing dimensions;
-    #   time must be present for all dimension combos;
-    #   - if dimension != project.dimension, then a project_dimension_mapping
-    #   table must be provided (for each mismatching dimension)??
-    #   Else: we need some way to verify the expected project dim mappings
+    # TODO: 
+    #   1. validate data matches dimensions specified in dataset config;
+    #   2. check that all required dimensions exist in the data or partitioning
+    #   3. check expected counts of things
+    #   4. check for required tables accounted for; 
+    #   5. any specific scaling factor checks?
 
 
 class DatasetConfig(DSGBaseModel):
     """Represents model dataset configurations"""
-    project_version: str = Field(
-        title="project_version",
-        description="project version",
-    )
     dataset_id: str = Field(
         title="dataset_id",
         description="dataset identifier",
@@ -117,7 +93,7 @@ class DatasetConfig(DSGBaseModel):
         title="path",
         description="path containing data",
     )
-    dimensions: List[Union[DatasetDimension, TimeDimension]] = Field(
+    dimensions: List[Union[Dimension, TimeDimension]] = Field(
         title="dimensions",
         description="dimensions defined by the dataset",
     )
@@ -129,8 +105,6 @@ class DatasetConfig(DSGBaseModel):
 
     # TODO: can we reuse this validator? Its taken from the
     #   project_config Diemnsions model
-    @validator('dimensions',
-               pre=True, each_item=True, always=True)
     def handle_dimension_union(cls, value):
         """
         Validate dimension type work around for pydantic Union bug
@@ -141,88 +115,12 @@ class DatasetConfig(DSGBaseModel):
         if value['type'] == DimensionType.TIME.value:
             val = TimeDimension(**value)
         else:
-            val = DatasetDimension(**value)
+            val = Dimension(**value)
         return val
 
-    # TODO: maybe this type of func lives outside of the pydantic config in
-    #   the wrapper?
-    def get_project_registry(project_version):
-        """Get the project registry associated with the file"""
-        # TODO: this needs to pull from S3 or some other central store
-        #   at some point
-        pr_path = f'./registry/projects/{project_version}.toml'
-        # pr = ProjectRegistry(**toml.load(pr_path))
-        pr = load_project_registry(pr_path)
-        return pr
-
-    @validator('project_version')
-    def check_active_project_registration(cls, project_version):
-        """
-        Validate that the project registration exists and that it
-        is not deprecated.
-        """
-        pr = cls.get_project_registry(project_version)
-        # # TODO using .value here because enum isn't saving value by default
-        # TODO: this does not work unless we use exception; ValueError results
-        #   in error w/ next validator on dataset_id"
-        if pr.status.value == 'Deprecated':
-            raise DSGBaseException(
-                'Project registration for project handle is deprecated')
-        return project_version
-
-    @validator('dataset_id')
-    def check_project_registry_for_dataset_id(cls, dataset_id, values):
-        """Check that dataset has not already been registered to project"""
-        # if 'project_version'
-        project_version = values['project_version']
-        pr = cls.get_project_registry(project_version)
-        dataset_ids = []
-        for dataset in pr.dataset_registries:
-            dataset_ids.append(dataset.dataset_id)
-        # check that dataset is expected by project
-        if dataset_id not in dataset_ids:
-            raise ValueError(
-                f'Dataset ID "{dataset_id}" is not expected by project')
-        # check that dataset has not already been registered to project
-        for dataset in pr.dataset_registries:
-            if dataset.dataset_id == dataset_id:
-                if dataset.status.value == 'Registered':
-                    raise ValueError(
-                        f'Dataset ID "{dataset_id}" is already registered to '
-                        f'project version "{project_version}"')
-        return dataset_id
-
-    # TODO: Need help here. Issues relate to enumeration not being stored as
-    #   value. Also, I don't know how to loop through these 3 fields properly.
-    #   I originally tried a validator for ('dataset_type', 'model_name',
-    #   'model_sector')
-    @root_validator(pre=True)
-    def check_for_project_config_expectations(cls, values):
-        """
-        Check dataset fields to project fields.
-        Fields to check: id, model_name, model, sectors
-        """
-        # TODO: Need help here. I tried to make this a validator function but
-        #   I was hitting issues with field names and the enumeration classes
-        #   not storing the values. This is the best I could do ATM.
-        dataset_id = values['dataset_id']
-        project_version = values['project_version']
-        pr = cls.get_project_registry(project_version)
-        pc = pr.project_config
-        for dataset in pc.input_datasets.datasets:
-            if dataset.dataset_id == dataset_id:
-                fields = ['dataset_type', 'model_name', 'model_sector']
-                for field in fields:
-                    data_val = values[field]
-                    pc_val = getattr(dataset, field)
-                    if isinstance(type(pc_val), Enum):
-                        pc_val = pc_val.value
-                    if pc_val != data_val:
-                        raise ValueError(
-                            f'Dataset {field}="{data_val}" but the project '
-                            f'config expects "{pc_val}".')
-        return values
-
+    # TODO: if local path provided, we want to upload to S3 and set the path 
+    #   here to S3 path
+    
     @validator("path")
     def check_path(cls, path):
         """Check dataset parquet path"""
@@ -257,27 +155,14 @@ class DatasetConfig(DSGBaseModel):
 
         return path
 
-
-# TODO where do this live? We can't put it in registry.py because
-#   the dataset config needs to reference the project registry ATM
-#   I can't put it in its own dataset_registry because it needs to reference
-#   the project registry
-#   @Dthom - help!
-def RegisterDataset(config_toml):
-    """
-    Register a dataset with a registered dsgrid project
-     given datast configuration toml.
-    """
-    # ----------------------------------
-    # TODO ACTION ITEMS
-    # assign a version
-    # make dataset regitry file
-    # add to project registry file and save it
-    # ----------------------------------
-
-    # validate dataset config
-    dataset_config = DatasetConfig(**toml.load(config_toml))
-    print("Dataset Config Validated")  # TODO: log message
-
-    project_config = dataset_config.get_project_registry(
-        dataset_config.project_version)
+    # TODO:
+    #   - check binning/partitioning / file size requirements?
+    #   - check unique names
+    #   - check unique files
+    #   - add similar validators as project_config Dimensions
+    # NOTE: project_config.Dimensions has a lot of the
+    #       validators we want here however, its unclear to me how we can
+    #       apply them in both classes because they are root valitors and
+    #       also because Dimensions includes project_dimension and
+    #       supplemental_dimensions which are not required at the dataset
+    #       level.
